+++ docs/examples/intro/HelloWorld.vue
@@ -1,0 +1,11 @@
+<script setup lang="ts">
+import { defineLayer } from 'vue-layerx'
+
+defineLayer({
+  props: { title: 'Hello World' },
+})
+</script>
+
+<template>
+  <p>你好，vue-layerx。</p>
+</template>