+++ docs/examples/intro/layers.ts
@@ -1,0 +1,4 @@
+import { ElDialog } from 'element-plus'
+import { createLayer } from 'vue-layerx'
+
+export const useDialog = createLayer(ElDialog)